Laying the ground for Plant Edit.
Add ID as a mapped field to PlantFormFields, as we need to receive it from PlantForm when onSubmit happens on edited item.

Add enableReinitialize to PlantForm as we change initialValues passed to either based on new Plant object or existing object from API.

Add handling of click on Plants row which now directs Plant edit.

Add route change tracking and reset initialValues.