UX improvements.
Show progress indicator during Plant Edit when plants are still loading.

Datepicker modifications:
Add calendar icon button to datepicker which activates it.
Make datepicker font size (and, as a consequence, the whole picker)
the same size as the rest of the application.

Fix issue in Firefox where text selected in custom select would be white.