Use Redux to manage categories state.
Turn off serializableCheck in middleware in order to prevent errors
(namely 'A non-serializable value was detected in the state' error, which
appears when we try to serialize an array of class-instantiated objects,
in our case -- categories).

Implement Categories reducer, actions and selector using createSlice.
This approach reduces (no pun intended) amount of usual Redux boilerplate
code and allows defining reducers in a convenient, mutable-like form
(they are then translated to their immutable counterparts).

Implement custom useCategories hook and utilize it in both withCategories
HOC (used on Plants) and CategoriesContainer component.